backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
twilio==8.10.0
google-generativeai==0.3.0
boto3==1.34.0
openai-whisper==20231117
ffmpeg-python==0.2.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-dateutil==2.8.2
stripe==7.6.0
email-validator==2.1.0
python-dotenv==1.0.0
python-magic==0.4.27
aiofiles==23.2.1
pillow==10.1.0
httpx==0.25.1